# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VaultX Crypto Investment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    put:
      operationId: updateUserProfile
      tags: [users]
      summary: Update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/change-password:
    post:
      operationId: changePassword
      tags: [users]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordBody"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /2fa/verify:
    post:
      operationId: verify2fa
      tags: [twofa]
      summary: Verify and enable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFAVerifyBody"
      responses:
        "200":
          description: 2FA enabled with backup codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TwoFAEnabled"
  /2fa/disable:
    post:
      operationId: disable2fa
      tags: [twofa]
      summary: Disable 2FA
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TwoFAVerifyBody"
      responses:
        "200":
          description: 2FA disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /wallet/deposit:
    post:
      operationId: createDeposit
      tags: [wallet]
      summary: Request a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositBody"
      responses:
        "201":
          description: Deposit initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/withdraw:
    post:
      operationId: createWithdrawal
      tags: [wallet]
      summary: Request a withdrawal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawBody"
      responses:
        "201":
          description: Withdrawal requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/transfer:
    post:
      operationId: createTransfer
      tags: [wallet]
      summary: Internal transfer to another user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferBody"
      responses:
        "201":
          description: Transfer completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /wallet/resolve-user:
    get:
      operationId: resolveTransferUser
      tags: [wallet]
      summary: Look up a user by username or user ID for transfer
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedUser"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        country:
          type: string
        avatarUrl:
          type: string
        referralCode:
          type: string
        kycStatus:
          type: string
          enum: [none, pending, approved, rejected]
        twoFaEnabled:
          type: boolean
        isAdmin:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, fullName, username, email, referralCode, kycStatus, twoFaEnabled, isAdmin, isVerified, createdAt]
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: string
        whatsapp:
          type: string
        country:
          type: string
        avatarUrl:
          type: string
    ChangePasswordBody:
      type: object
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
        confirmPassword:
          type: string
      required: [currentPassword, newPassword, confirmPassword]
...
    TwoFAVerifyBody:
      type: object
      properties:
        code:
          type: string
      required: [code]
    TwoFAEnabled:
      type: object
      properties:
        enabled:
          type: boolean
        backupCodes:
          type: array
          items:
            type: string
      required: [enabled, backupCodes]
...
    DepositBody:
      type: object
      properties:
        amount:
          type: number
        network:
          type: string
        txHash:
          type: string
      required: [amount, network]
    WithdrawBody:
      type: object
      properties:
        amount:
          type: number
        network:
          type: string
        address:
          type: string
      required: [amount, network, address]
    TransferBody:
      type: object
      properties:
        recipientQuery:
          type: string
        amount:
          type: number
        note:
          type: string
      required: [recipientQuery, amount]
    ResolvedUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        avatarUrl:
          type: string
      required: [id, username, fullName]
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [deposit, withdrawal, transfer, earning, investment, referral]
        amount:
          type: number
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        network:
          type: string
        txHash:
          type: string
        address:
          type: string
        note:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, amount, status, createdAt]
...